ETL with Kafka/requirements.txt
kafka-python
mysql-connector-python==8.0.31
